requirements.txt
beautifulsoup4==4.12.2
blinker==1.6.2
click==8.1.7
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
greenlet==2.0.2
gunicorn==21.2.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.1
psycopg2-binary==2.9.7
python-dotenv==1.0.0
soupsieve==2.4.1
SQLAlchemy==2.0.20
typing_extensions==4.7.1
Werkzeug==2.3.7
